GreyScale/requirements.txt
ipdb==0.13.13
ipython==8.12.0
pillow==10.2.0
pika===1.3.2
requests==2.29.0
